+++ fl_image_picker/example/app/web/flutter_service_worker.js
@@ -3,7 +3,7 @@ const MANIFEST = 'flutter-app-manifest';
 const TEMP = 'flutter-temp-cache';
 const CACHE_NAME = 'flutter-app-cache';
 
-const RESOURCES = {"flutter_bootstrap.js": "0863bc7467122e7a927c16fa54616fa8",
+const RESOURCES = {"flutter_bootstrap.js": "9d7a4a23f1a6a0c127f6c8cadc7d6d02",
 "version.json": "ff966ab969ba381b900e61629bfb9789",
 "index.html": "90cc34c76ba128594ba7ca03c740a12c",
 "/": "90cc34c76ba128594ba7ca03c740a12c",